:memo: :: (#176) move style to one file
