--- zamong/src/components/MainPage/globalstyle.ts
@@ -1,11 +1,0 @@
-import styled from "@emotion/styled";
-import { font } from "../../style/font";
-
-export const SectionTitle = styled.div`
-  font: ${font.headline3};
-  color: white;
-  margin-bottom: 28px;
-  display: flex;
-  justify-content: space-between;
-  align-items: flex-end;
-`;
