+++ zamong/src/components/MainPage/styles.ts
@@ -51,3 +51,12 @@ export const SectionContainer = styled.div`
   display: flex;
   flex-direction: column;
 `;
+
+export const SectionTitle = styled.div`
+  font: ${font.headline3};
+  color: white;
+  margin-bottom: 28px;
+  display: flex;
+  justify-content: space-between;
+  align-items: flex-end;
+`;
